requirements.txt
requests
azure-storage-blob
duckdb
polars
pyarrow